Updated a few names in the Vulkan folder, changed some object creation behaviour. Updated the Vulkan API to a beta version to experiment with the newly released VK_KHR_acceleration_structure and VK_KHR_ray_tracing_pipeline extentensions.

+++ OpenGL 3D Game/src/engine/platform/opengl/rendering/shaders/environmentMap/GLPrefilterShader.cpp
@@ -27,5 +27,5 @@ void GLPrefilterShader::UpdateUniforms(const Matrix4f& projection, const Matrix4
 	SetUniformi("environmentMap", 0);
 
 	SetUniformf("environmentMap", roughness);
-	SetUniformf("resolution", EnvironmentMapConfig::prefilterMapResolution);
+	SetUniformf("resolution", (float)EnvironmentMapConfig::prefilterMapResolution);
 }
